configs: New config in drivers/net/phy for v5.4-rc1
Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/net/phy/Kconfig (help): ADIN_PHY (tristate)
    Adds support for the Analog Devices Industrial Ethernet PHYs.
    Currently supports the:
    - ADIN1200 - Robust,Industrial, Low Power 10/100 Ethernet PHY
    - ADIN1300 - Robust,Industrial, Low Latency 10/100/1000 Gigabit
      Ethernet PHY

Cc: Petr Oros <poros@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>
Acked-by: Jiri Benc <jbenc@redhat.com>