[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_GPIO_LOGICVC:

 Say yes here to support GPIO functionality of the Xylon LogiCVC
 programmable logic block.

 Symbol: GPIO_LOGICVC [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:315
   Prompt: Xylon LogiCVC GPIO support
   Depends on: GPIOLIB [=y] && HAS_IOMEM [=y] && MFD_SYSCON [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Memory mapped GPIO drivers

---