[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SERIAL_8250_16550A_VARIANTS:

 The 8250 driver can probe for many variants of the venerable 16550A
 serial port. Doing so takes additional time at boot.

 On modern systems, especially those using serial only for a simple
 console, you can say N here.

 Symbol: SERIAL_8250_16550A_VARIANTS [=n]
 Type  : bool
 Defined at drivers/tty/serial/8250/Kconfig:63
   Prompt: Support for variants of the 16550A serial port
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_8250 [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Serial drivers
           -> 8250/16550 and compatible serial support (SERIAL_8250 [=y])

---

Prarit Bhargava requested that this be enabled: "Enabling this will make
virt systems boot faster.  We want this to be enabled for RHEL"

Cc: Prarit Bhargava <prarit@redhat.com>