[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HW_RANDOM_CCTRNG:

 Say 'Y' to enable the True Random Number Generator driver for the
 Arm TrustZone CryptoCell family of processors.
 Currently the CryptoCell 713 and 703 are supported.
 The driver is supported only in SoC where Trusted Execution
 Environment is not used.
 Choose 'M' to compile this driver as a module. The module
 will be called cctrng.
 If unsure, say 'N'.

 Symbol: HW_RANDOM_CCTRNG [=n]
 Type  : tristate
 Defined at drivers/char/hw_random/Kconfig:477
   Prompt: Arm CryptoCell True Random Number Generator support
   Depends on: HW_RANDOM [=y] && HAS_IOMEM [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Hardware Random Number Generator Core support (HW_RANDOM [=y])

---

Signed-off-by: CKI@GitLab <cki-project@redhat.com>