New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_INTEL_SCU_PCI:

 This driver is used to bridge the communications between kernel
 and SCU on some embedded Intel x86 platforms. It also creates
 devices that are connected to the SoC through the SCU.
 Platforms supported:
   Medfield
   Clovertrail
   Merrifield
   Broxton
   Apollo Lake

 Symbol: INTEL_SCU_PCI [=n]
 Type  : bool
 Defined at drivers/platform/x86/Kconfig:1354
   Prompt: Intel SCU PCI driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: INTEL_SCU [=n]
 Selected by [n]:
   - X86_INTEL_MID [=n] && X86_EXTENDED_PLATFORM [=y] && X86_PLATFORM_DEVICES [=y] && PCI [=y] && (X86_64 [=y] || PCI_GOANY [=n] && X86_32 [=n]) && X86_IO_APIC [=y]

Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>