Add dist-fedora-release target
This is a target that will create the rawhide branch, ark-latest.

It is mostly copied from the gitlab-ci.yml and turned into a
Makefile target for easier understanding and execution.

An optimization is to move some of the duplicate code to
git-tree-check and replace git-status with git-diff-index
to make it easier to detect a dirty tree.

This target will be called by the maintainer and the gitlab-ci.yml
file.  Currently unused.

No noticable changes.

Signed-off-by: Don Zickus <dzickus@redhat.com>