[redhat] New configs in drivers/acpi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ACPI_PLATFORM_PROFILE:

 This driver adds support for platform-profiles on platforms that
 support it.

 Platform-profiles can be used to control the platform behaviour. For
 example whether to operate in a lower power mode, in a higher
 power performance mode or between the two.

 This driver provides the sysfs interface and is used as the registration
 point for platform specific drivers.

 Which profiles are supported is determined on a per-platform basis and
 should be obtained from the platform specific driver.

 Symbol: ACPI_PLATFORM_PROFILE [=m]
 Type  : tristate
 Defined at drivers/acpi/Kconfig:329
   Prompt: ACPI Platform Profile Driver
   Depends on: ACPI [=y]
   Location:
     -> Power management and ACPI options
       -> ACPI (Advanced Configuration and Power Interface) Support (ACPI [=y])

---

Cc: Al Stone <ahs3@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>