[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_COMMON_CLK_AXI_CLKGEN:

 Support for the Analog Devices axi-clkgen pcore clock generator for Xilinx
 FPGAs. It is commonly used in Analog Devices' reference designs.

 Symbol: COMMON_CLK_AXI_CLKGEN [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:248
   Prompt: AXI clkgen driver
   Depends on: COMMON_CLK [=y] && (HAS_IOMEM [=y] || COMPILE_TEST [=n]) && OF [=y]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>