[redhat] New configs in drivers/vdpa
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_VP_VDPA:

 This kernel module bridges virtio PCI device to vDPA bus.

 Symbol: VP_VDPA [=n]
 Type  : tristate
 Defined at drivers/vdpa/Kconfig:63
   Prompt: Virtio PCI bridge vDPA driver
   Depends on: VDPA [=m] && PCI_MSI [=y]
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
 Selects: VIRTIO_PCI_LIB [=y]

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 - CONFIG_VP_VDPA enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>