[redhat] New configs in drivers/block
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ZRAM_DEF_COMP_842 [=n]
 Type  : bool
 Defined at drivers/block/zram/Kconfig:42
   Prompt: 842
   Depends on: <choice> && CRYPTO_842 [=m]
   Location:
     -> Device Drivers
       -> Block devices (BLK_DEV [=y])
         -> Compressed RAM block device support (ZRAM [=m])
           -> Default zram compressor (<choice> [=y])

---

 Symbol: ZRAM_DEF_COMP_LZO [=n]
 Type  : bool
 Defined at drivers/block/zram/Kconfig:34
   Prompt: lzo
   Depends on: <choice> && CRYPTO_LZO [=y]
   Location:
     -> Device Drivers
       -> Block devices (BLK_DEV [=y])
         -> Compressed RAM block device support (ZRAM [=m])
           -> Default zram compressor (<choice> [=y])

---

 Symbol: ZRAM_DEF_COMP_LZORLE [=y]
 Type  : bool
 Defined at drivers/block/zram/Kconfig:22
   Prompt: lzo-rle
   Depends on: <choice> && CRYPTO_LZO [=y]
   Location:
     -> Device Drivers
       -> Block devices (BLK_DEV [=y])
         -> Compressed RAM block device support (ZRAM [=m])
           -> Default zram compressor (<choice> [=y])

---

Cc: Ming Lei <ming.lei@redhat.com>
Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>