[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_MSM [=n]
 Type  : tristate
 Defined at drivers/pinctrl/qcom/Kconfig:4
   Prompt: Qualcomm core pin controller driver
   Depends on: PINCTRL [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && GPIOLIB [=y] && (QCOM_SCM [=n] || !QCOM_SCM [=n])
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
 Selects: PINMUX [=y] && PINCONF [=y] && GENERIC_PINCONF [=y] && GPIOLIB_IRQCHIP [=y] && IRQ_DOMAIN_HIERARCHY [=y] && IRQ_FASTEOI_HIERARCHY_HANDLERS [=n]

---

Cc: Brian Masney <bmasney@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>