redhat/kernel.spec.template: Add include_fedora and include_rhel variables
The centos-stream-9 (c9s) [1] repository contains files for both RHEL and
Fedora configs.  The Fedora configs will never be updated, and will
quickly be out-of-date with respect to upstream.  Having the files in
place causes confusion having the Fedora files causes for engineers and
contributors.

Add include_fedora and include_rhel variables.  These can be set in
Makefile.variables to include or exclude RHEL or Fedora-specific files.
These are enabled for ARK, and include_fedora can be disabled for c9s.

[1] https://gitlab.com/redhat/centos-stream/src/kernel/centos-stream-9/
Signed-off-by: Prarit Bhargava <prarit@redhat.com>