[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SERIAL_8250_PERICOM [=y]
 Type  : tristate
 Defined at drivers/tty/serial/8250/Kconfig:501
   Prompt: Support for Pericom and Acces I/O serial ports
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_8250 [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Character devices
           -> Enable TTY (TTY [=y])
             -> Serial drivers

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>