[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEGRA186_GPC_DMA [=n]
 Type  : tristate
 Defined at drivers/dma/Kconfig:632
   Prompt: NVIDIA Tegra GPC DMA support
   Depends on: DMADEVICES [=y] && (ARCH_TEGRA [=y] || COMPILE_TEST [=n]) && ARCH_DMA_ADDR_T_64BIT [=y] && IOMMU_API [=y]
   Location:
     Main menu
       -> Device Drivers
         -> DMA Engine support (DMADEVICES [=y])
 Selects: DMA_ENGINE [=y]

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>