[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: A64FX_DIAG [=n]
 Type  : bool
 Defined at drivers/soc/fujitsu/Kconfig:4
   Prompt: A64FX diag driver
   Depends on: ARM64 [=y] && ACPI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> SOC (System On Chip) specific Drivers
           -> fujitsu SoC drivers

---

 Symbol: QCOM_ICC_BWMON [=n]
 Type  : tristate
 Defined at drivers/soc/qcom/Kconfig:235
   Prompt: QCOM Interconnect Bandwidth Monitor driver
   Depends on: ARCH_QCOM [=y] || COMPILE_TEST [=n]
   Location:
     Main menu
       -> Device Drivers
         -> SOC (System On Chip) specific Drivers
           -> Qualcomm SoC drivers
 Selects: PM_OPP [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>