Turn on CONFIG_SQUASHFS_DECOMP_SINGLE in pending
Currently we have CONFIG_SQUASHFS_DECOMP_SINGLE turned off. Since 6.2 is
changing how DECOMP threading is managed in the config, this creates a
mismatch. This commit turns it on in pending, while a follow up MR will
set all of the values in a way that gives us the behavior we had before.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>