[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_ELKHARTLAKE [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1266
   Prompt: Intel Elkhart Lake PSE GPIO support
   Depends on: GPIOLIB [=y] && (X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> MFD GPIO expanders
           -> Intel Elkhart Lake PSE GPIO support (GPIO_ELKHARTLAKE [=n])
 Selects: GPIO_TANGIER [=n]

---

 Symbol: GPIO_FXL6408 [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1021
   Prompt: FXL6408 I2C GPIO expander
   Depends on: GPIOLIB [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> I2C GPIO expanders
           -> FXL6408 I2C GPIO expander (GPIO_FXL6408 [=n])
 Selects: GPIO_REGMAP [=n] && REGMAP_I2C [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>