[redhat] New configs in drivers/cdx
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CDX_BUS [=n]
 Type  : bool
 Defined at drivers/cdx/Kconfig:8
   Prompt: CDX Bus driver
   Depends on: OF [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> CDX Bus driver (CDX_BUS [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- rebased on 6.4.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>