[redhat] New configs in drivers/powercap
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_RAPL_TPMI [=n]
 Type  : tristate
 Defined at drivers/powercap/Kconfig:38
   Prompt: Intel RAPL Support via TPMI Interface
   Depends on: POWERCAP [=y] && X86 [=y] && INTEL_TPMI [=m]
   Location:
     -> Device Drivers
       -> Generic powercap sysfs driver (POWERCAP [=y])
         -> Intel RAPL Support via TPMI Interface (INTEL_RAPL_TPMI [=n])
 Selects: INTEL_RAPL_CORE [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_INTEL_RAPL_TPMI enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>