[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMPERE_ERRATUM_AC03_CPU_38 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:417
   Prompt: AmpereOne: AC03_CPU_38: Certain bits in the Virtualization Translation Control Register and Translation Control Registers do not follow RES0 semantics
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> AmpereOne: AC03_CPU_38: Certain bits in the Virtualization Translation Control Register and Translation Control Registers do not follow RES0 semantics (AMPERE_ERRATUM_AC03_CPU_38 [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>