[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MT7925E [=n]
 Type  : tristate
 Defined at drivers/net/wireless/mediatek/mt76/mt7925/Kconfig:7
   Prompt: MediaTek MT7925E (PCIe) support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_MEDIATEK [=y] && MAC80211 [=m] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> MediaTek devices (WLAN_VENDOR_MEDIATEK [=y])
             -> MediaTek MT7925E (PCIe) support (MT7925E [=n])
 Selects: MT7925_COMMON [=n]

---

 Symbol: MT7925U [=n]
 Type  : tristate
 Defined at drivers/net/wireless/mediatek/mt76/mt7925/Kconfig:19
   Prompt: MediaTek MT7925U (USB) support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_MEDIATEK [=y] && MAC80211 [=m] && USB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> MediaTek devices (WLAN_VENDOR_MEDIATEK [=y])
             -> MediaTek MT7925U (USB) support (MT7925U [=n])
 Selects: MT792x_USB [=n] && MT7925_COMMON [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>