[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_IMX93_MIPI_DSI [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/bridge/imx/Kconfig:52
   Prompt: Freescale i.MX93 specific extensions for Synopsys DW MIPI DSI
   Depends on: HAS_IOMEM [=y] && DRM [=y] && DRM_BRIDGE [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n]) && OF [=y] && COMMON_CLK [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Interface Bridges
           -> Freescale i.MX93 specific extensions for Synopsys DW MIPI DSI (DRM_IMX93_MIPI_DSI [=n])
 Selects: DRM_DW_MIPI_DSI [=n] && GENERIC_PHY [=y] && GENERIC_PHY_MIPI_DPHY [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>